refactor : ConsoleApplication 기능을 Controller로 분리
