+++ src/main/java/team/marco/vouchermanagementsystem/service/BlacklistService.java
@@ -13,10 +13,7 @@ public class BlacklistService {
         this.blacklistRepository = blacklistRepository;
     }
 
-    public List<String> getBlacklist() {
-        return blacklistRepository.findAll()
-                .stream()
-                .map(User::getInfo)
-                .toList();
+    public List<User> getBlacklist() {
+        return blacklistRepository.findAll();
     }
 }
